requirements.txt
# Django Configuration
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-environ>=0.11.2
psycopg2-binary>=2.9.9
redis>=5.0.0
celery>=5.3.0
requests>=2.31.0
gunicorn>=21.2.0
whitenoise>=6.6.0
